+++ package.json
@@ -47,7 +47,7 @@
     "nyc": "^10.0.0",
     "request": "^2.79.0",
     "standard": "^8.0.0",
-    "tap": "^10.0.0",
+    "tap": "^10.0.2",
     "tape": "^4.6.2"
   }
 }
